+++ src/test/java/ru/netology/stats/StatsServiceTest.java
@@ -11,8 +11,8 @@ public class StatsServiceTest {
 
         long[] salesStat = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
-        int expectedAmount = 180;
-        int actualAmount = service.calcAmountSales(salesStat);
+        long expectedAmount = 180;
+        long actualAmount = service.calcAmountSales(salesStat);
 
         Assertions.assertEquals(expectedAmount, actualAmount);
     }
@@ -22,8 +22,8 @@ public class StatsServiceTest {
         StatsService service = new StatsService();
 
         long[] salesStat = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
-        int expectedAverage = 15;
-        int actualAmount = service.calcAverageAmountSales(salesStat);
+        long expectedAverage = 15;
+        long actualAmount = service.calcAverageAmountSales(salesStat);
 
         Assertions.assertEquals(expectedAverage, actualAmount);
     }
